+++ src/headers/image_hdr.rs
@@ -5,10 +5,9 @@ use std::fmt::Display;
 use std::fs::File;
 use std::str::FromStr;
 
-use crate::error::NitfError;
 use crate::headers::NitfSegmentHeader;
-use crate::types::{NitfField, Security, ExtendedSubheader};
-
+use crate::types::{ExtendedSubheader, NitfField, Security};
+use crate::{NitfError, NitfResult};
 /// Metadata for Image Segment subheader
 #[derive(Default, Clone, Debug, Eq, PartialEq)]
 pub struct ImageHeader {
@@ -93,7 +92,7 @@ pub struct ImageHeader {
     /// Image Extended Subheader Overflow
     pub ixsofl: NitfField<u16>,
     /// Image Extended Subheader Data
-    pub ixshd: ExtendedSubheader
+    pub ixshd: ExtendedSubheader,
 }
 
 /// Band metadata
@@ -241,90 +240,91 @@ pub enum Mode {
 
 // FUNCTIONS
 /// Helper function for parsing bands
-fn read_bands(reader: &mut File, n_band: u32) -> Vec<Band> {
+fn read_bands(reader: &mut File, n_band: u32) -> NitfResult<Vec<Band>> {
     let mut bands: Vec<Band> = vec![Band::default(); n_band as usize];
     for band in &mut bands {
-        band.irepband.read(reader, 2u8);
-        band.isubcat.read(reader, 6u8);
-        band.ifc.read(reader, 1u8);
-        band.imflt.read(reader, 3u8);
-        band.nluts.read(reader, 1u8);
+        band.irepband.read(reader, 2u8, "IREPBAND")?;
+        band.isubcat.read(reader, 6u8, "ISUBCAT")?;
+        band.ifc.read(reader, 1u8, "IFC")?;
+        band.imflt.read(reader, 3u8, "IMFLT")?;
+        band.nluts.read(reader, 1u8, "NLUTS")?;
         if band.nluts.val != 0 {
-            band.nelut.read(reader, 5u8);
+            band.nelut.read(reader, 5u8, "NELUT")?;
             for _ in 0..band.nelut.val {
                 let mut lut: NitfField<u8> = NitfField::default();
-                lut.read(reader, 1u8);
+                lut.read(reader, 1u8, "LUDT")?;
                 band.lutd.push(lut);
             }
         }
     }
-    bands
+    Ok(bands)
 }
 
 // TRAIT IMPLEMENTATIONS
 impl NitfSegmentHeader for ImageHeader {
-    fn read(&mut self, reader: &mut File) {
-        self.im.read(reader, 2u8);
-        self.iid1.read(reader, 10u8);
-        self.idatim.read(reader, 14u8);
-        self.tgtid.read(reader, 17u8);
-        self.iid2.read(reader, 80u8);
-        self.security.read(reader);
-        self.encryp.read(reader, 1u8);
-        self.isorce.read(reader, 42u8);
-        self.nrows.read(reader, 8u8);
-        self.ncols.read(reader, 8u8);
-        self.pvtype.read(reader, 3u8);
-        self.irep.read(reader, 8u8);
-        self.icat.read(reader, 8u8);
-        self.abpp.read(reader, 2u8);
-        self.pjust.read(reader, 1u8);
-        self.icords.read(reader, 1u8);
+    fn read(&mut self, reader: &mut File) -> NitfResult<()> {
+        self.im.read(reader, 2u8, "IM")?;
+        self.iid1.read(reader, 10u8, "IID1")?;
+        self.idatim.read(reader, 14u8, "IDATIM")?;
+        self.tgtid.read(reader, 17u8, "TGTID")?;
+        self.iid2.read(reader, 80u8, "IID2")?;
+        self.security.read(reader)?;
+        self.encryp.read(reader, 1u8, "ENCRYP")?;
+        self.isorce.read(reader, 42u8, "ISORCE")?;
+        self.nrows.read(reader, 8u8, "NROWS")?;
+        self.ncols.read(reader, 8u8, "NCOLS")?;
+        self.pvtype.read(reader, 3u8, "PVTYPE")?;
+        self.irep.read(reader, 8u8, "IREP")?;
+        self.icat.read(reader, 8u8, "ICAT")?;
+        self.abpp.read(reader, 2u8, "ABPP")?;
+        self.pjust.read(reader, 1u8, "PJUST")?;
+        self.icords.read(reader, 1u8, "ICORDS")?;
         for _ in 0..4 {
             let mut geoloc: NitfField<String> = NitfField::default();
-            geoloc.read(reader, 15u8);
+            geoloc.read(reader, 15u8, "READ")?;
             self.igeolo.push(geoloc);
         }
-        self.nicom.read(reader, 1u8);
+        self.nicom.read(reader, 1u8, "NICOM")?;
         for _ in 0..self.nicom.val {
             let mut comment: NitfField<String> = NitfField::default();
-            comment.read(reader, 80u8);
+            comment.read(reader, 80u8, "READ")?;
             self.icoms.push(comment);
         }
 
-        self.ic.read(reader, 2u8);
-        self.nbands.read(reader, 1u8);
+        self.ic.read(reader, 2u8, "IC")?;
+        self.nbands.read(reader, 1u8, "NBANDS")?;
         // If NBANDS = 0, use XBANDS
         if self.nbands.val != 0 {
-            self.bands = read_bands(reader, self.nbands.val as u32)
+            self.bands = read_bands(reader, self.nbands.val as u32)?;
         } else {
-            self.xbands.read(reader, 5u8);
-            self.bands = read_bands(reader, self.xbands.val)
+            self.xbands.read(reader, 5u8, "XBANDS")?;
+            self.bands = read_bands(reader, self.xbands.val)?;
         }
-        self.isync.read(reader, 1u8);
-        self.imode.read(reader, 1u8);
-        self.nbpr.read(reader, 4u8);
-        self.nbpc.read(reader, 4u8);
-        self.nppbh.read(reader, 4u8);
-        self.nppbv.read(reader, 4u8);
-        self.nbpp.read(reader, 2u8);
-        self.idlvl.read(reader, 3u8);
-        self.ialvl.read(reader, 3u8);
-        self.iloc.read(reader, 10u8);
-        self.imag.read(reader, 4u8);
-        self.udidl.read(reader, 5u8);
+        self.isync.read(reader, 1u8, "ISYNC")?;
+        self.imode.read(reader, 1u8, "IMODE")?;
+        self.nbpr.read(reader, 4u8, "NBPR")?;
+        self.nbpc.read(reader, 4u8, "NBPC")?;
+        self.nppbh.read(reader, 4u8, "NPPBH")?;
+        self.nppbv.read(reader, 4u8, "NPPBV")?;
+        self.nbpp.read(reader, 2u8, "NBPP")?;
+        self.idlvl.read(reader, 3u8, "IDLVL")?;
+        self.ialvl.read(reader, 3u8, "IALVL")?;
+        self.iloc.read(reader, 10u8, "ILOC")?;
+        self.imag.read(reader, 4u8, "IMAG")?;
+        self.udidl.read(reader, 5u8, "UDIDL")?;
         let udi_data_length = self.udidl.val;
         if udi_data_length != 0 {
-            self.udofl.read(reader, 3u8);
-            self.udid.read(reader, udi_data_length - 3);
+            self.udofl.read(reader, 3u8, "UDOFL")?;
+            self.udid.read(reader, udi_data_length - 3, "UDID")?;
         }
-        self.ixshdl.read(reader, 5u8);
+        self.ixshdl.read(reader, 5u8, "IXSHDL")?;
         let ixsh_data_length = self.ixshdl.val;
         if ixsh_data_length != 0 {
-            self.ixsofl.read(reader, 3u8);
-            self.ixshd.read(reader, (ixsh_data_length - 3) as usize);
-            
+            self.ixsofl.read(reader, 3u8, "IXSOFL")?;
+            self.ixshd
+                .read(reader, (ixsh_data_length - 3) as usize, "IXSHD")?;
         }
+        Ok(())
     }
 }
 impl Display for ImageHeader {
@@ -373,8 +373,7 @@ impl Display for ImageHeader {
         out_str += format!("UDOFL: {}, ", self.udofl).as_ref();
         out_str += format!("UDID: {}, ", self.udid).as_ref();
         out_str += format!("IXSHDL: {}, ", self.ixshdl).as_ref();
-        if self.ixshdl.val != 0
-        {
+        if self.ixshdl.val != 0 {
             out_str += format!("IXSOFL: {}, ", self.ixsofl).as_ref();
             out_str += format!("IXSHD: {}", self.ixshd).as_ref();
         }
@@ -405,7 +404,7 @@ impl FromStr for PixelValueType {
             "SI" => Ok(Self::SI),
             "R" => Ok(Self::R),
             "C" => Ok(Self::C),
-            _ => Err(NitfError::FieldError),
+            _ => Err(NitfError::EnumError("PixelValueType")),
         }
     }
 }
@@ -422,7 +421,7 @@ impl FromStr for ImageRepresentation {
             "POLAR" => Ok(Self::POLAR),
             "VPH" => Ok(Self::VPH),
             "YCbCr601" => Ok(Self::YCbCr601),
-            _ => Err(NitfError::FieldError),
+            _ => Err(NitfError::EnumError("ImageRepresentation")),
         }
     }
 }
@@ -432,7 +431,7 @@ impl FromStr for PixelJustification {
         match s {
             "R" => Ok(Self::R),
             "L" => Ok(Self::L),
-            _ => Err(NitfError::FieldError),
+            _ => Err(NitfError::EnumError("PixelJustification")),
         }
     }
 }
@@ -447,7 +446,7 @@ impl FromStr for CoordinateRepresentation {
             "P" => Ok(Self::P),
             "G" => Ok(Self::G),
             "D" => Ok(Self::D),
-            _ => Err(NitfError::FieldError),
+            _ => Err(NitfError::EnumError("CoordinateRepresentation")),
         }
     }
 }
@@ -472,7 +471,7 @@ impl FromStr for Compression {
             "M6" => Ok(Self::M6),
             "M7" => Ok(Self::M7),
             "M8" => Ok(Self::M8),
-            _ => Err(NitfError::FieldError),
+            _ => Err(NitfError::EnumError("Compression")),
         }
     }
 }
@@ -484,7 +483,7 @@ impl FromStr for Mode {
             "P" => Ok(Self::P),
             "R" => Ok(Self::R),
             "S" => Ok(Self::S),
-            _ => Err(NitfError::FieldError),
+            _ => Err(NitfError::EnumError("Mode")),
         }
     }
 }