+++ src/headers/reserved_extension_hdr.rs
@@ -3,8 +3,8 @@ use std::fmt::Display;
 use std::fs::File;
 
 use crate::headers::NitfSegmentHeader;
-use crate::types::{NitfField, Security, ExtendedSubheader};
-
+use crate::types::{ExtendedSubheader, NitfField, Security};
+use crate::NitfResult;
 /// Metadata for Reserved Extension Segment
 #[derive(Default, Clone, Debug, Eq, PartialEq)]
 pub struct ReservedExtensionHeader {
@@ -34,14 +34,16 @@ impl Display for ReservedExtensionHeader {
     }
 }
 impl NitfSegmentHeader for ReservedExtensionHeader {
-    fn read(&mut self, reader: &mut File) {
-        self.re.read(reader, 2u8);
-        self.resid.read(reader, 25u8);
-        self.resver.read(reader, 2u8);
-        self.security.read(reader);
-        self.resshl.read(reader, 4u8);
+    fn read(&mut self, reader: &mut File) -> NitfResult<()> {
+        self.re.read(reader, 2u8, "RE")?;
+        self.resid.read(reader, 25u8, "RESID")?;
+        self.resver.read(reader, 2u8, "RESVER")?;
+        self.security.read(reader)?;
+        self.resshl.read(reader, 4u8, "RESSHL")?;
         if self.resshl.val != 0 {
-            self.resshf.read(reader, self.resshl.val as usize);
+            self.resshf
+                .read(reader, self.resshl.val as usize, "RESSHF")?;
         }
+        Ok(())
     }
 }