+++ src/types.rs
@@ -1,10 +1,11 @@
 //! Common types use throughout
+use log::{trace, warn};
 use std::fmt::{Debug, Display};
-use std::str::FromStr;
-use std::io::{Read, Seek};
 use std::fs::File;
+use std::io::{Read, Seek};
+use std::str::FromStr;
 
-use crate::error::NitfError;
+use crate::{NitfError, NitfResult};
 
 /// Lowest level object for file parsing
 #[derive(Default, Clone, Debug, Eq, PartialEq)]
@@ -146,26 +147,49 @@ pub enum ClassificationReason {
 /// Use Default implementation
 impl<V> NitfField<V>
 where
-    V: FromStr + Debug,
+    V: FromStr + Debug + Default,
     <V as FromStr>::Err: Debug,
 {
     /// Read the specified number of bytes and parse the value of a given field
-    pub fn read<T: Sized + Into<u64>>(&mut self, reader: &mut File, n_bytes: T) {
+    pub fn read<T: Sized + Into<u64>>(
+        &mut self,
+        reader: &mut File,
+        n_bytes: T,
+        field_name: &str,
+    ) -> NitfResult<()> {
         self.length = n_bytes.into();
         self.bytes = vec![0; self.length as usize];
-        self.offset = reader.stream_position().unwrap();
-        reader.read_exact(&mut self.bytes).unwrap();
-        let result = String::from_utf8(self.bytes.to_vec());
-        match result {
+
+        // Crash if something goes wrong with the cursor
+        self.offset = reader
+            .stream_position()
+            .or(Err(NitfError::Fatal(field_name.to_string())))?;
+
+        // Crash if there is an error reading the bytes
+        reader
+            .read_exact(&mut self.bytes)
+            .or(Err(NitfError::Fatal(field_name.to_string())))?;
+
+        // Try to read the bytes to a string
+        match String::from_utf8(self.bytes.to_vec()) {
+            // If it's ok, trim and try to parse to enum/native representation
             Ok(str) => {
                 self.string = str.trim().to_string();
-                self.val = self.string.parse().unwrap();
+
+                // Warn and assign a default value if error parsing
+                self.val = self.string.parse().unwrap_or_else(|_| {
+                    warn!("Non-fatal error parsing {}", field_name);
+                    V::default()
+                });
             }
-            Err(err) => {
+
+            Err(_) => {
                 self.string = String::from("Error parsing string");
-                println!("{}", err)
+                warn!("Failed to parse {field_name} from bytes: {:?}", self.bytes);
             }
         }
+        trace!("{:?}", self.val);
+        Ok(())
     }
 }
 impl<V: FromStr + Debug> Display for NitfField<V> {
@@ -174,23 +198,24 @@ impl<V: FromStr + Debug> Display for NitfField<V> {
     }
 }
 impl Security {
-    pub fn read(&mut self, reader: &mut File) {
-        self.clas.read(reader, 1u8);
-        self.clsy.read(reader, 2u8);
-        self.code.read(reader, 11u8);
-        self.ctlh.read(reader, 2u8);
-        self.rel.read(reader, 20u8);
-        self.dctp.read(reader, 2u8);
-        self.dcdt.read(reader, 8u8);
-        self.dcxm.read(reader, 4u8);
-        self.dg.read(reader, 1u8);
-        self.dgdt.read(reader, 8u8);
-        self.cltx.read(reader, 43u8);
-        self.catp.read(reader, 1u8);
-        self.caut.read(reader, 40u8);
-        self.crsn.read(reader, 1u8);
-        self.srdt.read(reader, 8u8);
-        self.ctln.read(reader, 15u8);
+    pub fn read(&mut self, reader: &mut File) -> NitfResult<()> {
+        self.clas.read(reader, 1u8, "CLAS")?;
+        self.clsy.read(reader, 2u8, "CLSY")?;
+        self.code.read(reader, 11u8, "CODE")?;
+        self.ctlh.read(reader, 2u8, "CTLH")?;
+        self.rel.read(reader, 20u8, "REL")?;
+        self.dctp.read(reader, 2u8, "DCTP")?;
+        self.dcdt.read(reader, 8u8, "DCDT")?;
+        self.dcxm.read(reader, 4u8, "DCXM")?;
+        self.dg.read(reader, 1u8, "DG")?;
+        self.dgdt.read(reader, 8u8, "DGDT")?;
+        self.cltx.read(reader, 43u8, "CLTX")?;
+        self.catp.read(reader, 1u8, "CATP")?;
+        self.caut.read(reader, 40u8, "CAUT")?;
+        self.crsn.read(reader, 1u8, "CRSN")?;
+        self.srdt.read(reader, 8u8, "SRDT")?;
+        self.ctln.read(reader, 15u8, "CTLN")?;
+        Ok(())
     }
 }
 impl Display for Security {
@@ -224,7 +249,7 @@ impl FromStr for Classification {
             "S" => Ok(Self::S),
             "C" => Ok(Self::C),
             "R" => Ok(Self::R),
-            _ => Err(NitfError::FieldError),
+            _ => Err(NitfError::EnumError("Classification")),
         }
     }
 }
@@ -239,7 +264,7 @@ impl FromStr for DeclassificationType {
             "GE" => Ok(Self::GE),
             "O" => Ok(Self::O),
             "X" => Ok(Self::X),
-            _ => Err(NitfError::FieldError),
+            _ => Err(NitfError::EnumError("DeclassificationType")),
         }
     }
 }
@@ -248,26 +273,29 @@ impl FromStr for DeclassificationExemption {
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         match s {
             "" => Ok(Self::DEFAULT),
-            "X1"   => Ok(Self::VALID),  // DOD 5200.01-V1, 4-201b(1)
-            "X2"   => Ok(Self::VALID),  // DOD 5200.01-V1, 4-201b(2)
-            "X3"   => Ok(Self::VALID),  // DOD 5200.01-V1, 4-201b(3)
-            "X4"   => Ok(Self::VALID),  // DOD 5200.01-V1, 4-201b(4)
-            "X5"   => Ok(Self::VALID),  // DOD 5200.01-V1, 4-201b(5)
-            "X6"   => Ok(Self::VALID),  // DOD 5200.01-V1, 4-201b(6)
-            "X7"   => Ok(Self::VALID),  // DOD 5200.01-V1, 4-201b(7)
-            "X8"   => Ok(Self::VALID),  // DOD 5200.01-V1, 4-201b(8)
-            "25X1" => Ok(Self::VALID),  // DOD 5200.01-V1, 4-301b(1)
-            "25X2" => Ok(Self::VALID),  // DOD 5200.01-V1, 4-301b(2)
-            "25X3" => Ok(Self::VALID),  // DOD 5200.01-V1, 4-301b(3)
-            "25X4" => Ok(Self::VALID),  // DOD 5200.01-V1, 4-301b(4)
-            "25X5" => Ok(Self::VALID),  // DOD 5200.01-V1, 4-301b(5)
-            "25X6" => Ok(Self::VALID),  // DOD 5200.01-V1, 4-301b(6)
-            "25X7" => Ok(Self::VALID),  // DOD 5200.01-V1, 4-301b(7)
-            "25X8" => Ok(Self::VALID),  // DOD 5200.01-V1, 4-301b(8)
-            "25X9" => Ok(Self::VALID),  // DOD 5200.01-V1, 4-301b(9)
+            "X1" => Ok(Self::VALID),   // DOD 5200.01-V1, 4-201b(1)
+            "X2" => Ok(Self::VALID),   // DOD 5200.01-V1, 4-201b(2)
+            "X3" => Ok(Self::VALID),   // DOD 5200.01-V1, 4-201b(3)
+            "X4" => Ok(Self::VALID),   // DOD 5200.01-V1, 4-201b(4)
+            "X5" => Ok(Self::VALID),   // DOD 5200.01-V1, 4-201b(5)
+            "X6" => Ok(Self::VALID),   // DOD 5200.01-V1, 4-201b(6)
+            "X7" => Ok(Self::VALID),   // DOD 5200.01-V1, 4-201b(7)
+            "X8" => Ok(Self::VALID),   // DOD 5200.01-V1, 4-201b(8)
+            "25X1" => Ok(Self::VALID), // DOD 5200.01-V1, 4-301b(1)
+            "25X2" => Ok(Self::VALID), // DOD 5200.01-V1, 4-301b(2)
+            "25X3" => Ok(Self::VALID), // DOD 5200.01-V1, 4-301b(3)
+            "25X4" => Ok(Self::VALID), // DOD 5200.01-V1, 4-301b(4)
+            "25X5" => Ok(Self::VALID), // DOD 5200.01-V1, 4-301b(5)
+            "25X6" => Ok(Self::VALID), // DOD 5200.01-V1, 4-301b(6)
+            "25X7" => Ok(Self::VALID), // DOD 5200.01-V1, 4-301b(7)
+            "25X8" => Ok(Self::VALID), // DOD 5200.01-V1, 4-301b(8)
+            "25X9" => Ok(Self::VALID), // DOD 5200.01-V1, 4-301b(9)
             "DN10" => Ok(Self::VALID),
-            "DNI"  => Ok(Self::VALID),
-            _      => Ok(Self::UNRECOGNIZED)
+            "DNI" => Ok(Self::VALID),
+            _ => {
+                warn!("Unrecognized DeclassificationExemption");
+                Ok(Self::UNRECOGNIZED)
+            }
         }
     }
 }
@@ -279,7 +307,7 @@ impl FromStr for Downgrade {
             "S" => Ok(Self::S),
             "C" => Ok(Self::C),
             "R" => Ok(Self::R),
-            _ => Err(NitfError::FieldError),
+            _ => Err(NitfError::EnumError("Downgrade")),
         }
     }
 }
@@ -291,7 +319,7 @@ impl FromStr for ClassificationAuthorityType {
             "O" => Ok(Self::O),
             "D" => Ok(Self::D),
             "M" => Ok(Self::M),
-            _ => Err(NitfError::FieldError),
+            _ => Err(NitfError::EnumError("ClassificationAuthorityType")),
         }
     }
 }
@@ -308,7 +336,7 @@ impl FromStr for ClassificationReason {
             "F" => Ok(Self::VALID),
             "G" => Ok(Self::VALID),
             "H" => Ok(Self::VALID),
-            _ => Err(NitfError::FieldError),
+            _ => Err(NitfError::EnumError("ClassificationReason")),
         }
     }
 }
@@ -319,20 +347,21 @@ pub struct ExtendedSubheader {
     pub tre: Vec<u8>,
     /// Length of subheader
     pub size: usize,
-
 }
 impl ExtendedSubheader {
-    pub fn read(&mut self, reader: &mut File, n_bytes: usize) {
+    pub fn read(&mut self, reader: &mut File, n_bytes: usize, name: &str) -> NitfResult<()> {
         self.size = n_bytes;
         self.tre = vec![0; n_bytes];
-        reader.read_exact(self.tre.as_mut_slice()).unwrap();
-
+        reader
+            .read_exact(self.tre.as_mut_slice())
+            .or(Err(NitfError::Fatal(name.to_string())))?;
+        Ok(())
     }
 }
 impl Display for ExtendedSubheader {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         // Can't copy vector directly, convert to slice, then clone into new vector
-        let out_str = String::from_utf8(self.tre.to_vec()).unwrap();
+        let out_str = String::from_utf8(self.tre.to_vec()).or(Err(std::fmt::Error))?;
         write!(f, "[{out_str}]")
     }
 }